Move the common help function to utils folder
